created numberic domain object

+++ src/main/scala-2.11/relationalClustering/representation/domain/Domain.scala
@@ -7,7 +7,7 @@ class Domain(val name: String) {
 
   var elements = collection.mutable.Set[String]()
 
-  def size = { elements.size }
+  def size = { elements.size.toDouble }
 
   def hasElement(element: String) = { elements.contains(element) }
